back/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
decorator==5.2.1
Django==5.2.4
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
et_xmlfile==2.0.0
executing==2.2.0
idna==3.11
ipython==9.4.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.4.0
openpyxl==3.1.5
pandas==2.3.3
parso==0.8.4
pillow==11.3.0
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
